Double -> Floatに変換し余計なbracketを削除

+++ app/Main.hs
@@ -4,17 +4,17 @@ module Main where
 main :: IO ()
 main = do
     putStrLn "現在の勤務時間?"
-    currentOperatingTime :: Double <- readLn
+    currentOperatingTime :: Float <- readLn
     putStrLn "今月の最低稼働時間?"
-    minimumUpTime :: Double <- readLn
+    minimumUpTime :: Float <- readLn
     putStrLn "今月の残り稼働日数?"
     remainingOperatingDay :: Int <- readLn
     print $ minimumUpTimePerDay remainingOperatingDay $ remainingTime currentOperatingTime minimumUpTime
 
 -- 最低稼働時間を満たすための残り稼働時間を算出
-remainingTime :: Double -> Double -> Double
+remainingTime :: Float -> Float -> Float
 remainingTime currentTime minimumUpTime = minimumUpTime - currentTime
 
 -- 1日あたりの最低稼働時間を算出
-minimumUpTimePerDay :: Int -> Double -> Double
-minimumUpTimePerDay remainingDay remainingTime = remainingTime / (realToFrac (remainingDay :: Int))
+minimumUpTimePerDay :: Int -> Float -> Float
+minimumUpTimePerDay remainingDay remainingTime = remainingTime / realToFrac (remainingDay :: Int)
